+++ test/all.js
@@ -8,6 +8,7 @@ exports['test multiple listeners'] = require('./multiple-listeners');
 exports['test methods'] = require('./methods');
 exports['test thenable'] = require('./thenable');
 exports['test spread'] = require('./spread');
+exports['test allResolved'] = require("./all-resolved");
 exports['test node'] = require('./node');
 
 exports['test GH issue 9'] = require('./issue/9');